% Generated by roxygen2 (4.0.2): do not edit by hand
\name{block_hook}
\alias{block_hook}
\alias{hook_html}
\alias{hook_latex}
\alias{hook_markdown_latest}
\alias{hook_markdown_v1}
\alias{hook_markdown_v2}
\title{Chunk Block building hooks.}
\usage{
block_hook(out.format = "markdown")

hook_html(before, options, envir, name = "literal")

hook_latex(before, options, envir, name = "literal")

hook_markdown_v1(before, options, envir, name = "literal")

hook_markdown_v2(before, options, envir, name = "literal")

hook_markdown_latest(before, options, envir, name = "literal")
}
\arguments{
\item{out.format}{Name of out.format as given by
 \code{opts_knit$get("out.format")}}

\item{before}{If \code{TRUE} chunk code has not been executed.}

\item{options}{Current knitr chunk options.}

\item{envir}{Environment in which code chunk is evaluated.}

\item{name}{Name of the hook as \code{$set} in knitr.}
}
\value{
\code{character vector} of built block if \code{before}
is \code{TRUE}.

block_hook returns the \code{hook_} function for out.format.
}
\description{
Block building knitr hooks.

Dispatches the block building hook to be set in knitr.
}

