% Generated by roxygen2 (4.0.0): do not edit by hand
\name{chop_params}
\alias{chop_params}
\title{Removes all trailing chunk parameters beginning at the
last appearance of \code{name=} or any \code{opts.label} matching
\code{"name.*"}.}
\usage{
chop_params(params.src, name = "literal")
}
\arguments{
\item{params.src}{string as returned by \code{opts_chunk$get("params.src")}.}

\item{name}{Name of the hook as \code{$set} in knitr.}
}
\value{
\code{character vector} of the modified parameters.
}
\description{
Removes all trailing chunk parameters beginning at the
last appearance of \code{name=} or any \code{opts.label} matching
\code{"name.*"}.
}
\note{
Be sure that the provided hook name is unique enough to not
  match unintentded paramters.
}

