% Generated by roxygen2 (4.0.0): do not edit by hand
\name{get_code}
\alias{get_code}
\title{Knitr expanded source code for the chunk.}
\usage{
get_code(options, name)
}
\arguments{
\item{options}{Chunk paramters as provided by the knitr hook call.}

\item{name}{Name of the hook as \code{$set} in knitr.}
}
\value{
\code{character vector} of the modified parameters.
}
\description{
Knitr expanded source code for the chunk.
}
\note{
Honors \code{tidy} as knitr does.
}

