% Generated by roxygen2 (4.0.2): do not edit by hand
\name{kast_on}
\alias{kast_on}
\title{Set the knitr hook.}
\usage{
kast_on(name = "literal", templates = TRUE, include_styles = TRUE)
}
\arguments{
\item{name}{Name string to use for chunk option, option template
and .opt prefixes.}

\item{templates}{If \code{TRUE} option templates will be enabled.}

\item{include_styles}{include supported styling files after body.}
}
\value{
invisible \code{TRUE} on success.
}
\description{
Set the knitr hook.
}
\note{
templates provided include: \cr
"name" eval=FALSE, echo=FALSE \cr
"name-eval" echo=FALSE \cr
"name-reuse" eval=FALSE, echo=FALSE, literal.opts=list(empty=TRUE) \cr
"name-quote" echo=FALSE, results="asis", indent="> "
}

