% Generated by roxygen2 (4.0.0): do not edit by hand
\name{opt_template}
\alias{opt_template}
\alias{opt_templates}
\alias{opt_templates_html}
\alias{opt_templates_latex}
\alias{opt_templates_markdown}
\title{knitr option templates}
\usage{
opt_templates(out.format, name = "literal")

opt_templates_html()

opt_templates_latex()

opt_templates_markdown()
}
\arguments{
\item{out.format}{Name of out.format as given by
\code{opts_knit$get("out.format")}}
}
\value{
block_hook returns the \code{hook_} function for out.format.
}
\description{
knitr chunk option templates to be used with
  the knitr \code{opts.label} parameter.

Dispatches for the option templates to be set in knitr.
}
\details{
Calls \code{knitr::opts_template$set()} for each
  builtin option template for the matching \code{out.format}.
}

